[BTAT-2248] Refactored Income Source related models to incomeSourceDetails package
