+++ test/connectors/IncomeSourceDetailsConnectorSpec.scala
@@ -19,7 +19,7 @@ package connectors
 import assets.IncomeSourceDetailsTestConstants._
 import assets.BaseTestConstants._
 import mocks.MockHttp
-import models.IncomeSourceDetailsError
+import models.incomeSourceDetails.IncomeSourceDetailsError
 import play.mvc.Http.Status
 import uk.gov.hmrc.http.logging.Authorization
 import uk.gov.hmrc.http.{HeaderCarrier, HttpResponse}
